Enforce permissions for project or chart modifications

+++ cloudman/projman/api.py
@@ -36,6 +36,20 @@ class PMService(object):
         """
         return self._context
 
+    def has_permissions(self, scopes, obj=None):
+        if not isinstance(scopes, list):
+            scope = [scopes]
+        return self.context.user.has_perms(scope, obj)
+
+    def check_permissions(self, scopes, obj=None):
+        if not self.has_permissions(scopes, obj):
+            self.raise_no_permissions(scopes)
+
+    def raise_no_permissions(self, scopes):
+        raise PermissionError(
+            "Object does not exist or you do not have permissions to "
+            "perform '%s'" % (scopes,))
+
 
 class ProjManAPI(PMService):
 
@@ -58,34 +72,51 @@ class PMProjectService(PMService):
     def __init__(self, context):
         super(PMProjectService, self).__init__(context)
 
+    def to_api_object(self, project):
+        # Remap the returned django model's delete method to the API method
+        # This is just a lazy alternative to writing an actual wrapper around
+        # the django object.
+        project.delete = lambda: self.delete(project.id)
+        return self.add_child_services(project)
+
     def add_child_services(self, project):
         project.service = self
         project.charts = PMProjectChartService(self.context, project)
         return project
 
     def list(self):
-        return list(map(self.add_child_services,
-                        models.CMProject.objects.all()))
+        return list(map(
+            self.to_api_object,
+            (proj for proj in models.CMProject.objects.all()
+             if self.has_permissions('projects.view_project', proj))))
 
     def get(self, project_id):
-        return self.add_child_services(
-            models.CMProject.objects.get(id=project_id))
+        obj = models.CMProject.objects.get(id=project_id)
+        self.check_permissions('projects.view_project', obj)
+        return self.to_api_object(obj)
 
     def create(self, name):
+        self.check_permissions('projects.add_project')
         obj = models.CMProject.objects.create(
-            name=name)
-        project = self.add_child_services(obj)
+            name=name, owner=self.context.user)
+        project = self.to_api_object(obj)
         return project
 
     def delete(self, project_id):
         obj = models.CMProject.objects.get(id=project_id)
         if obj:
+            self.check_permissions('projects.delete_project', obj)
             obj.delete()
+        else:
+            self.raise_no_permissions('projects.delete_project')
 
     def find(self, name):
         try:
-            return self.add_child_services(
-                models.CMProject.objects.get(name=name))
+            obj = models.CMProject.objects.get(name=name)
+            if self.has_permissions('projects.view_project', obj):
+                return self.to_api_object(obj)
+            else:
+                return None
         except models.CMProject.DoesNotExist:
             return None
 
@@ -101,25 +132,40 @@ class PMProjectChartService(PMService):
 
     def _to_proj_chart(self, chart):
         chart.project = self.project
+        # Remap the helm API's delete method to the project chart API method
+        # This is just a lazy alternative to writing an actual wrapper around
+        # the HelmChart object.
+        chart.delete = lambda: self.delete(chart.id)
         return chart
 
     def list(self):
         return [self._to_proj_chart(chart) for chart
                 in self._get_helmsman_api().charts.list()
-                if chart.namespace == self.project.name]
+                if chart.namespace == self.project.name and
+                self.has_permissions('charts.view_chart', chart)]
 
     def get(self, chart_id):
         chart = self._get_helmsman_api().charts.get(chart_id)
+        self.check_permissions('charts.view_chart', chart)
         return (self._to_proj_chart(chart)
                 if chart and chart.namespace == self.project.name else None)
 
     def create(self, repo_name, chart_name, release_name=None, version=None,
                values=None):
+        self.check_permissions('charts.add_chart')
         return self._to_proj_chart(self._get_helmsman_api().charts.create(
             repo_name, chart_name, self.project.name, release_name, version,
             values))
 
+    def update(self, chart, values):
+        self.check_permissions('charts.change_chart', chart)
+        updated_chart = self._get_helmsman_api().charts.update(chart, values)
+        return self._to_proj_chart(updated_chart)
+
     def delete(self, chart_id):
         obj = self.get(chart_id)
         if obj:
-            obj.delete()
+            self.check_permissions('charts.delete_chart', obj)
+            self._get_helmsman_api().charts.delete(obj)
+        else:
+            self.raise_no_permissions('charts.delete_chart')
